refactor: reorganize code to enable reusing the router
