+++ pkg/controllers/cleanup.go
@@ -3,7 +3,7 @@ package controllers
 import (
 	"net/http"
 
-	"github.com/envelope-zero/backend/internal/database"
+	"github.com/envelope-zero/backend/pkg/database"
 	"github.com/envelope-zero/backend/pkg/httperrors"
 	"github.com/envelope-zero/backend/pkg/models"
 	"github.com/gin-gonic/gin"
