+++ pkg/database/database.go
@@ -2,8 +2,11 @@ package database
 
 import (
 	"fmt"
+	"os"
+	"path/filepath"
 	"time"
 
+	"github.com/glebarez/sqlite"
 	gorm_zerolog "github.com/wei840222/gorm-zerolog"
 	"gorm.io/gorm"
 )
@@ -43,6 +46,24 @@ func ConnectDatabase(dialector func(string) gorm.Dialector, dsn string) error {
 	sqlDB.SetMaxOpenConns(1)
 
 	DB = db
+
+	return nil
+}
+
+// Database sets up the database.
+func Database() error {
+	// Create data directory
+	dataDir := filepath.Join(".", "data")
+	err := os.MkdirAll(dataDir, os.ModePerm)
+	if err != nil {
+		return err
+	}
+
+	// Connect to the database
+	err = ConnectDatabase(sqlite.Open, "data/gorm.db?_pragma=foreign_keys(1)")
+	if err != nil {
+		return err
+	}
 
 	return nil
 }
