+++ pkg/models/database_test.go
@@ -1,0 +1,11 @@
+package models_test
+
+import (
+	"github.com/envelope-zero/backend/pkg/models"
+	"github.com/stretchr/testify/assert"
+)
+
+func (suite *TestSuiteEnv) TestMigrateDatabase() {
+	err := models.MigrateDatabase()
+	assert.Nil(suite.T(), err)
+}
