+++ pkg/models/envelope.go
@@ -3,7 +3,7 @@ package models
 import (
 	"time"
 
-	"github.com/envelope-zero/backend/internal/database"
+	"github.com/envelope-zero/backend/pkg/database"
 	"github.com/google/uuid"
 	"github.com/shopspring/decimal"
 	"gorm.io/gorm"
