+++ pkg/router/router_test.go
@@ -7,8 +7,8 @@ import (
 	"os"
 	"testing"
 
-	"github.com/envelope-zero/backend/internal/router"
-	"github.com/envelope-zero/backend/pkg/test"
+	"github.com/envelope-zero/backend/pkg/router"
+	"github.com/envelope-zero/backend/test"
 	"github.com/gin-gonic/gin"
 	"github.com/stretchr/testify/assert"
 )
